RavenDB-18299 Corax - RAW type for indexing.
