+++ src/Corax/IndexEntryBuilder.cs
@@ -32,7 +32,7 @@ namespace Corax
         
         Tuple = 1 << 2,
         List = 1 << 3,
-        Blittable = 1 << 4,
+        Raw = 1 << 4,
         Invalid = 1 << 6,
     }
 
@@ -128,30 +128,25 @@ namespace Corax
             _dataIndex += length + value.Length;
         }
         
-        public void Write(int field, Span<byte> blittableBinary, Span<byte> blittableAsText)
+        public void WriteRaw(int field, ReadOnlySpan<byte> binaryValue)
         {
             //STRUCT
-            //<type><size_of_binary><size_of_text><binary><text>
+            //<type><size_of_binary><binary>
             Debug.Assert(field < _knownFields.Count);
             Debug.Assert(_knownFieldsLocations[field] == Invalid);
             
-            if (blittableBinary.Length == 0)
+            if (binaryValue.Length == 0)
                 return;
 
             int dataLocation = _dataIndex;
             // Write known field.
             _knownFieldsLocations[field] = dataLocation | unchecked((int)0x80000000);
-            dataLocation += VariableSizeEncoding.Write(_buffer, (byte)IndexEntryFieldType.Blittable, dataLocation);
-            dataLocation += VariableSizeEncoding.Write(_buffer, blittableBinary.Length, dataLocation);
-            dataLocation += VariableSizeEncoding.Write(_buffer, blittableAsText.Length, dataLocation);
-            
-            
-            blittableBinary.CopyTo(_buffer.Slice(dataLocation));
-            dataLocation += blittableBinary.Length;
-            
-            blittableAsText.CopyTo(_buffer.Slice(dataLocation));
-            dataLocation += blittableAsText.Length;
+            dataLocation += VariableSizeEncoding.Write(_buffer, (byte)IndexEntryFieldType.Raw, dataLocation);
+            dataLocation += VariableSizeEncoding.Write(_buffer, binaryValue.Length, dataLocation);
 
+            binaryValue.CopyTo(_buffer.Slice(dataLocation));
+            dataLocation += binaryValue.Length;
+            
             _dataIndex = dataLocation;
         }
 
@@ -836,6 +831,12 @@ namespace Corax
 
                     stringLength = VariableSizeEncoding.Read<int>(_buffer, out length, spanTableOffset);
                 }
+                else if ((type & IndexEntryFieldType.Raw) != 0)
+                {
+                    stringLength = VariableSizeEncoding.Read<int>(_buffer, out int readOffset, intOffset);
+                    intOffset += readOffset;
+                    type = IndexEntryFieldType.Raw;
+                }
                 else if ((type & IndexEntryFieldType.Tuple) != 0)
                 {
                     VariableSizeEncoding.Read<long>(_buffer, out length, intOffset); // Skip
@@ -857,32 +858,7 @@ namespace Corax
             value = _buffer.Slice(intOffset, stringLength);            
             return true;
         }
-
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public bool Read(int field, out Span<byte> blittableBinary, out Span<byte> blittableText, int intOffset = Invalid)
-        {
-            if (intOffset == Invalid)
-            {
-                (intOffset, var isTyped) =  GetMetadataFieldLocation(_buffer, field);
-                if (intOffset is Invalid)
-                {
-                    blittableBinary = Span<byte>.Empty;
-                    blittableText = Span<byte>.Empty;
-                    return false;
-                }
-            }
-
-            intOffset += sizeof(IndexEntryFieldType);
-            var blittableBinaryLength = VariableSizeEncoding.Read<int>(_buffer, out int lengthSize, intOffset);
-            intOffset += lengthSize;
-            var blittableTextLength = VariableSizeEncoding.Read<int>(_buffer, out lengthSize, intOffset);
-            intOffset += lengthSize;
-
-            blittableBinary = _buffer.Slice(intOffset, blittableBinaryLength);
-            blittableText = _buffer.Slice(intOffset + blittableBinaryLength, blittableTextLength);
-            return true;
-        }
-
+        
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Read(int field, out long longValue, out double doubleValue, out Span<byte> sequenceValue)
         {
