+++ src/Corax/IndexWriter.cs
@@ -30,126 +30,7 @@ namespace Corax
         Small = 1,
         Set = 2
     }
-
-
-    public class IndexFieldsMapping : IEnumerable<IndexFieldBinding>
-    {
-        public static readonly IndexFieldsMapping Instance = new IndexFieldsMapping(null);
-
-        private readonly ByteStringContext _context;
-        private readonly Dictionary<Slice, IndexFieldBinding> _fields;
-        private readonly Dictionary<int, IndexFieldBinding> _fieldsById;
-        private readonly List<IndexFieldBinding> _fieldsList;
-        public int Count => _fieldsById.Count;
-        public Analyzer DefaultAnalyzer;
-
-        public IndexFieldsMapping(ByteStringContext context)
-        {
-            _context = context;
-            _fields = new Dictionary<Slice, IndexFieldBinding>(SliceComparer.Instance);
-            _fieldsById = new Dictionary<int, IndexFieldBinding>();
-            _fieldsList = new List<IndexFieldBinding>();
-        }
-
-        public IndexFieldsMapping AddBinding(int fieldId, Slice fieldName, Analyzer analyzer = null, bool notApplyAnalyzer = false, bool hasSuggestion = false)
-        {
-            if (!_fieldsById.TryGetValue(fieldId, out var storedAnalyzer))
-            {
-                var binding = new IndexFieldBinding(fieldId, fieldName, analyzer, notApplyAnalyzer, hasSuggestion);
-                _fields[fieldName] = binding;
-                _fieldsById[fieldId] = binding;
-                _fieldsList.Add(binding);
-            }
-            else
-            {
-                Debug.Assert(analyzer == storedAnalyzer.Analyzer);
-            }
-
-            return this;
-        }
-
-        public void UpdateAnalyzersInBindings(IndexFieldsMapping analyzers)
-        {
-            foreach (var mapping in analyzers.GetEnumerator())
-            {
-                if (TryGetByFieldId(mapping.FieldId, out var binding) == true)
-                {
-                    binding.Analyzer = mapping.Analyzer;
-                }
-            }
-
-            foreach (var ifb in CollectionsMarshal.AsSpan(_fieldsList))
-            {
-                if (ifb.NotApplyAnalyzer)
-                    continue;
-
-                ifb.Analyzer ??= analyzers.DefaultAnalyzer;
-            }
-        }
-
-        public IndexFieldBinding GetByFieldId(int fieldId)
-        {
-            return _fieldsById[fieldId];
-        }
-
-        public bool TryGetByFieldId(int fieldId, out IndexFieldBinding binding)
-        {
-            return _fieldsById.TryGetValue(fieldId, out binding);
-        }
-
-        public IndexFieldBinding GetByFieldName(string fieldName)
-        {
-            // This method is a convenience method that should not be used in high performance sections of the code.
-            using var _ = Slice.From(_context, fieldName, out var str);
-            return _fields[str];
-        }
-
-        public bool TryGetByFieldName(string fieldName, out IndexFieldBinding binding)
-        {
-            // This method is a convenience method that should not be used in high performance sections of the code.
-            using var _ = Slice.From(_context, fieldName, out var str);
-            return _fields.TryGetValue(str, out binding);
-        }
-
-        public IndexFieldBinding GetByFieldName(Slice fieldName)
-        {
-            return _fields[fieldName];
-        }
-
-        public bool TryGetByFieldName(Slice fieldName, out IndexFieldBinding binding)
-        {
-            return _fields.TryGetValue(fieldName, out binding);
-        }
-
-        public IEnumerator<IndexFieldBinding> GetEnumerator()
-        {
-            return _fieldsList.GetEnumerator();
-        }
-
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return _fieldsList.GetEnumerator();
-        }
-    }
-
-    public class IndexFieldBinding
-    {
-        public readonly int FieldId;
-        public readonly Slice FieldName;
-        public Corax.Analyzer Analyzer;
-        public readonly bool NotApplyAnalyzer;
-        public readonly bool HasSuggestions;
-
-        public IndexFieldBinding(int fieldId, Slice fieldName, Analyzer analyzer = null, bool notApplyAnalyzer = false, bool hasSuggestions = false)
-        {
-            FieldId = fieldId;
-            FieldName = fieldName;
-            Analyzer = analyzer;
-            NotApplyAnalyzer = notApplyAnalyzer;
-            HasSuggestions = hasSuggestions;
-        }
-    }
-
+    
     public class IndexWriter : IDisposable // single threaded, controlled by caller
     {
         public const int MaxTermLength = 1024;
@@ -304,15 +185,13 @@ namespace Corax
             else if (fieldType.HasFlag(IndexEntryFieldType.Tuple) ||
                      fieldType.HasFlag(IndexEntryFieldType.Invalid) == false)
             {
-                Span<byte> value;
-                if (fieldType.HasFlag(IndexEntryFieldType.Blittable))
+                if (fieldType.HasFlag(IndexEntryFieldType.Raw))
                 {
-                    entryReader.Read(tokenField, out Span<byte> _, out value);
+                    return;
                 }
-                else
-                {
-                    entryReader.Read(tokenField, out value);
-                }
+                
+                entryReader.Read(tokenField, out var value);
+                
                 var words = new Span<byte>(tempWordsSpace, bufferSize);
                 var tokens = new Span<Token>(tempTokenSpace, tokenSize);
                 analyzer.Execute(value, ref words, ref tokens);
@@ -433,13 +312,10 @@ namespace Corax
             }
             else if (fieldType.HasFlag(IndexEntryFieldType.Tuple) || fieldType.HasFlag(IndexEntryFieldType.Invalid) == false)
             {
-                Span<byte> value;
-                if (fieldType.HasFlag(IndexEntryFieldType.Blittable))
-                    entryReader.Read(tokenField, out Span<byte> _, out value, intOffset);
-                else
-                    entryReader.Read(tokenField, out value);
-
-
+                if (fieldType.HasFlag(IndexEntryFieldType.Raw))
+                    return;
+                
+                entryReader.Read(tokenField, out var value);
                 using var _ = Slice.From(context, value, ByteStringType.Mutable, out var slice);
                 if (field.TryGetValue(slice, out var term) == false)
                 {
@@ -529,11 +405,10 @@ namespace Corax
                     }
                     else
                     {
-                        Span<byte> termValue;
-                        if (fieldType.HasFlag(IndexEntryFieldType.Blittable))
-                            entryReader.Read(fieldId, out Span<byte> _, out termValue);
-                        else
-                            entryReader.Read(fieldId, out termValue);
+                        if (fieldType.HasFlag(IndexEntryFieldType.Raw))
+                            continue;
+                        
+                        entryReader.Read(fieldId, out var termValue);
 
                         if (analyzer is null)
                         {
